refactored some files around
